requirements.txt
apache-beam==2.48.0
apache-flink==1.20.1
apache-flink-libraries==1.20.1
avro-python3==1.10.2
certifi==2025.1.31
charset-normalizer==3.4.1
cloudpickle==2.2.1
crcmod==1.7
dill==0.3.1.1
dnspython==2.7.0
docopt==0.6.2
fastavro==1.10.0
fasteners==0.19
find_libpython==0.4.0
grpcio==1.70.0
hdfs==2.7.3
httplib2==0.22.0
idna==3.10
kafka-python==2.0.3
numpy==1.24.4
objsize==0.6.1
orjson==3.10.15
pandas==2.2.3
pemja==0.4.1
proto-plus==1.26.0
protobuf==4.23.4
py4j==0.10.9.7
pyarrow==11.0.0
pydot==1.4.2
pymongo==4.11.1
pyparsing==3.2.1
pyspark==3.5.4
pyflink
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
regex==2024.11.6
requests==2.32.3
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
six==1.17.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
zstandard==0.23.0
apache-airflow~=2.10.5
apache-airflow[postgres]
psycopg2-binary
